+++ misc/examples/bitsets/bits.c
@@ -9,18 +9,18 @@ int main(void)
         cbits_drop(&set),
         cbits_drop(&s2)
     ){
-        printf("count %lld, %lld\n", cbits_count(&set), cbits_size(&set));
+        printf("count %d, %d\n", (int)cbits_count(&set), (int)cbits_size(&set));
         cbits s1 = cbits_from("1110100110111");
         char buf[256];
         cbits_to_str(&s1, buf, 0, 255);
-        printf("buf: %s: %lld\n", buf, cbits_count(&s1));
+        printf("buf: %s: %d\n", buf, (int)cbits_count(&s1));
         cbits_drop(&s1);
 
         cbits_reset(&set, 9);
         cbits_resize(&set, 43, false);
         printf(" str: %s\n", cbits_to_str(&set, buf, 0, 255));
 
-        printf("%4lld: ", cbits_size(&set));
+        printf("%4d: ", (int)cbits_size(&set));
         c_forrange (i, cbits_size(&set))
             printf("%d", cbits_test(&set, i));
         puts("");
@@ -30,12 +30,12 @@ int main(void)
         cbits_resize(&set, 93, false);
         cbits_resize(&set, 102, true);
         cbits_set_value(&set, 99, false);
-        printf("%4lld: ", cbits_size(&set));
+        printf("%4d: ", (int)cbits_size(&set));
         c_forrange (i, cbits_size(&set))
             printf("%d", cbits_test(&set, i));
 
         puts("\nIterate:");
-        printf("%4lld: ", cbits_size(&set));
+        printf("%4d: ", (int)cbits_size(&set));
         c_forrange (i, cbits_size(&set))
             printf("%d", cbits_test(&set, i));
         puts("");
@@ -58,7 +58,7 @@ int main(void)
         puts("");
 
         cbits_set_all(&set, false);
-        printf("%4lld: ", cbits_size(&set));
+        printf("%4d: ", (int)cbits_size(&set));
         c_forrange (i, cbits_size(&set))
             printf("%d", cbits_test(&set, i));
         puts("");
